+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.16.0 - 2022-03-21
 
 - Add Clojure icon [Thanks freshhops](https://github.com/BeardedBear/bearded-icons/pull/57)
+- Add Terraform icon [Thanks nahlian](https://github.com/BeardedBear/bearded-icons/pull/54)
 
 
 ## 0.15.0 - 2021-11-14